rag-python/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
jaydebeapi
jpype1
